RavenDB-13111 Reordering sharded/unsharded tests
